changed integer to bigint in sql table turns out that you cant use INTEGER for big numbers in sql, you have to use BIGINT relates #31

+++ database_build/db_build.sql
@@ -6,7 +6,7 @@ CREATE TABLE users (
   id          SERIAL        PRIMARY KEY,
   firstname   VARCHAR(30)   NOT NULL,
   lastname    VARCHAR(50),
-  facebook_id INTEGER       UNIQUE,
+  facebook_id BIGINT       UNIQUE,
   postcode    VARCHAR(10)
 );
